feat: extract global feature search
